+++ algonquian-real-estate/plugins/algq-marketplace/assets/js/algq-marketplace-admin.js
@@ -1,3 +1,9 @@
+(function ($) {
+  'use strict';
+  $(function () {
+    $('.algq-marketplace-admin').attr('data-algq-admin-ready', 'true');
+  });
+})(jQuery);
 (function () {
     'use strict';